+++ Assets/Scripts/ScriptableObjects/Generator.cs
@@ -32,28 +32,28 @@ namespace ScriptableObjects
         public Type _type; 
         [HideInInspector] public IOperator.State _state { get; set; }
         
-        public int _level { get; set; }
-        public int _production;
+        public ulong _level { get; set; }
+        public ulong _production;
         public Resource.Type _resource { get; set; }
         public Resource.Type _costResource { get; set; }
-        public int _levelCost { get; set; }
-        public int _purchaseCost { get; set; }
+        public ulong _levelCost { get; set; }
+        public ulong _purchaseCost { get; set; }
         public bool _requiresGenerator { get; set; }
         public Type _requiredGenerator { get; set; }
         public bool _requiresModifier { get; set; }
         public Modifier.Type _requiredModifier { get; set; }
-        public int _requiredLevel { get; set; }
+        public ulong _requiredLevel { get; set; }
         public bool isGenerator { get; set; }
 
         public void AddLevel()
         {
             _level++; 
         }
-        public void SetLevel(int level)
+        public void SetLevel(ulong level)
         {
             _level = level;
         }
-        public int GetLevel()
+        public ulong GetLevel()
         {
             return _level;
         }
