+++ Assets/Scripts/ScriptableObjects/GeneratorData.cs
@@ -11,14 +11,14 @@ namespace ScriptableObjects
         public Generator.Type _type;
         public Resource.Type _resource;
         public Resource.Type _costResource;
-        public int _purchaseCost;
-        public int _levelCost;
-        public int _production;
+        public ulong _purchaseCost;
+        public ulong _levelCost;
+        public ulong _production;
         public bool _requiresGenerator;
         public Generator.Type _requiredGenerator;
         public bool _requiresModifier;
         public Modifier.Type _requiredModifier;
-        public int _requiredLevel;
+        public ulong _requiredLevel;
 
         public Generator.Type GetGeneratorType()
         {
