+++ Assets/Scripts/ScriptableObjects/ModifierData.cs
@@ -10,15 +10,15 @@ namespace ScriptableObjects
         
         public Modifier.Type _type;
         public Resource.Type _costResource;
-        public int _purchaseCost;
-        public int _levelCost;
+        public ulong _purchaseCost;
+        public ulong _levelCost;
         public bool _requiresGenerator;
         public Generator.Type _requiredGenerator;
         public bool _requiresModifier;
         public Modifier.Type _requiredModifier;
-        public int _requiredLevel;
+        public ulong _requiredLevel;
         public float _creditPercentage = 0f;
-        public int _happinessCost = 0;
+        public ulong _happinessCost = 0;
         public float _levelPricePercentage = 0;
 
         public Modifier.Type GetModifierType()
