+++ Assets/Scripts/ScriptableObjects/Resource.cs
@@ -15,10 +15,10 @@ namespace ScriptableObjects
 
 
         public Type _type;
-        public int _amount; 
-        public int _newGame;
-        public int _devGame;
-        public int _minBound;    
-        public int _maxBound;
+        public ulong _amount; 
+        public ulong _newGame;
+        public ulong _devGame;
+        public ulong _minBound;    
+        public ulong _maxBound;
     }
 }
